requirements.txt
flask==2.3.3
flask-login==0.6.2
pymongo==4.5.0
flask-wtf==1.1.1
werkzeug==2.3.7
python-dotenv==1.0.0
dnspython==2.4.2
flask-bcrypt==1.0.1
flask-mail==0.9.1
gunicorn==21.2.0
pymongo[srv]
dnspython
